IMP l10n_it_riba_commission tests (#486)
