+++ l10n_it_ricevute_bancarie/__manifest__.py
@@ -9,7 +9,7 @@
 
 {
     'name': "Ricevute Bancarie",
-    'version': "10.0.1.0.2",
+    'version': "10.0.1.0.3",
     'author': "Odoo Community Association (OCA)",
     'category': "Accounting & Finance",
     'website': "https://odoo-community.org/",
