+++ l10n_it_ricevute_bancarie/tests/__init__.py
@@ -6,3 +6,4 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 from . import test_riba
+from . import riba_common
